Separated hit components into different files; moved constants to constants.js;
