+++ frontend/src/constants/constants.js
@@ -1,2 +1,9 @@
 export const BACKEND_URL = "https://api.projectias.neera.ai";
 export const GA_MEASUREMENT_ID = "G-QDM8PF4XX9";
+
+export const MAINS_URL = `${BACKEND_URL}/search_pyq`;
+export const PRELIMS_URL = `${BACKEND_URL}/search_prelims`;
+export const DNS_URL = `${BACKEND_URL}/search_dns`;
+export const CONTENT_URL = `${BACKEND_URL}/search_content`;
+
+export const LOG_URL = `${BACKEND_URL}/log`;
